+++ scripts/macros/modern.mjs
@@ -1,33 +1,33 @@
 // ** Class Features
 // Barbarian
-export {aspectOfTheWilds as 'aspect-of-the-wilds'} from './modern/classFeatures/barbarian/wildHeart/aspectOfTheWilds.mjs';
-export {brutalStrike as 'brutal-strike', forcefulBlow} from './modern/classFeatures/barbarian/brutalStrike.mjs';
-export {extendRage, keepRage} from './modern/classFeatures/barbarian/rage.mjs';
-export {improvedBrutalStrike as 'improved-brutal-strike', sunderingBlow} from './modern/classFeatures/barbarian/improvedBrutalStrike.mjs';
-export {indomitableMight as 'indomitable-might'} from './modern/classFeatures/barbarian/imdomitableMight.mjs';
 export {intimidatingPresence as 'intimidating-presence'} from './modern/classFeatures/barbarian/berserker/intimidatingPresence.mjs';
+export {brutalStrike as 'brutal-strike', forcefulBlow} from './modern/classFeatures/barbarian/brutalStrike.mjs';
+export {indomitableMight as 'indomitable-might'} from './modern/classFeatures/barbarian/imdomitableMight.mjs';
+export {improvedBrutalStrike as 'improved-brutal-strike', sunderingBlow} from './modern/classFeatures/barbarian/improvedBrutalStrike.mjs';
+export {unarguableWizardry as 'unarguable-wizardry'} from './modern/classFeatures/barbarian/muscleWizard/unarguableWizardry.mjs';
 export {persistentRage as 'persistent-rage'} from './modern/classFeatures/barbarian/persistentRage.mjs';
 export {primalKnowldege as 'primal-knowledge', primalKnowledgeEffect} from './modern/classFeatures/barbarian/primalKnowledge.mjs';
-export {unarguableWizardry as 'unarguable-wizardry'} from './modern/classFeatures/barbarian/muscleWizard/unarguableWizardry.mjs';
+export {extendRage, keepRage} from './modern/classFeatures/barbarian/rage.mjs';
+export {aspectOfTheWilds as 'aspect-of-the-wilds'} from './modern/classFeatures/barbarian/wildHeart/aspectOfTheWilds.mjs';
 // Bard
 export {countercharm} from './modern/classFeatures/bard/countercharm.mjs';
 export {fontOfInspiration as 'font-of-inspiration'} from './modern/classFeatures/bard/fontOfInspiration.mjs';
 // Cleric
-export {blessingOfTheTrickster as 'blessing-of-the-trickster'} from './modern/classFeatures/cleric/trickery/blessingOfTheTrickster.mjs';
 export {channelDivinity as 'channel-divinity|class:cleric', divineSparkDamage, turnUndead} from './modern/classFeatures/cleric/channelDivinity.mjs';
-export {discipleOfLife as 'disciple-of-life'} from './modern/classFeatures/cleric/life/discipleOfLife.mjs';
 export {divineIntervention as 'divine-intervention'} from './modern/classFeatures/cleric/divineIntervention.mjs';
 export {divineStrike as 'blessed-strikes-divine-strike'} from './modern/classFeatures/cleric/divineStrike.mjs';
 export {greaterDivineIntervention as 'greater-divine-intervention', greaterDivineInterventionRest as 'greater-divine-intervention-rest'} from './modern/classFeatures/cleric/greaterDivineIntervention.mjs';
 export {improvedPotentSpellcasting as 'improved-blessed-strikes'} from './modern/classFeatures/cleric/improvedPotentSpellcasting.mjs';
+export {discipleOfLife as 'disciple-of-life'} from './modern/classFeatures/cleric/life/discipleOfLife.mjs';
 export {improvedWardingFlare as 'improved-warding-flare'} from './modern/classFeatures/cleric/light/improvedWardingFlare.mjs';
 export {wardingFlare as 'warding-flare'} from './modern/classFeatures/cleric/light/wardingFlare.mjs';
+export {blessingOfTheTrickster as 'blessing-of-the-trickster'} from './modern/classFeatures/cleric/trickery/blessingOfTheTrickster.mjs';
 // Druid
 export {archdruid} from './modern/classFeatures/druid/archdruid.mjs';
+export {lunarForm as 'lunar-form'} from './modern/classFeatures/druid/moon/lunarForm.mjs';
 export {primalStrike as 'elemental-fury-primal-strike'} from './modern/classFeatures/druid/primalStrike.mjs';
 export {wildResurgence as 'wild-resurgence'} from './modern/classFeatures/druid/wildResurgence.mjs';
 export {wildShape as 'wild-shape'} from './modern/classFeatures/druid/wildShape.mjs';
-export {lunarForm as 'lunar-form'} from './modern/classFeatures/druid/moon/lunarForm.mjs';
 // Fighter
 export {heroicWarrior as 'heroic-warrior'} from './modern/classFeatures/fighter/heroicWarrior.mjs';
 // Monk
@@ -37,21 +37,29 @@ export {perfectFocus as 'perfect-focus'} from './modern/classFeatures/monk/perfe
 // Paladin
 export {layOnHands as 'lay-on-hands'} from './modern/classFeatures/paladin/layOnHands.mjs';
 // Ranger
+export {exceptionalTraining as 'exceptional-training'} from './modern/classFeatures/ranger/beastMaster/exceptionalTraining.mjs';
+export {bestialFury as 'bestial-fury', primalCompanion as 'primal-companion'} from './modern/classFeatures/ranger/beastMaster/primalCompanion.mjs';
+export {shareSpells as 'share-spells'} from './modern/classFeatures/ranger/beastMaster/shareSpells.mjs';
+export {favoredEnemy as 'favored-enemy'} from './modern/classFeatures/ranger/favoredEnemy.mjs';
+export {mistyWanderer as 'misty-wanderer'} from './modern/classFeatures/ranger/feyWanderer/mistyWanderer.mjs';
 export {foeSlayer as 'foe-slayer'} from './modern/classFeatures/ranger/foeSlayer.mjs';
-export {relentlessHunter as 'relentless-hunter'} from './modern/classFeatures/ranger/relentlessHunter.mjs';
+export {dreadAmbusher as 'dread-ambusher'} from './modern/classFeatures/ranger/gloomStalker/dreadAmbusher.mjs';
+export {shadowyDodge as 'shadowy-dodge'} from './modern/classFeatures/ranger/gloomStalker/shadowyDodge.mjs';
 export {stalkersFlurry as 'stalkers-flurry'} from './modern/classFeatures/ranger/gloomStalker/stalkersFlurry.mjs';
+export {preciseHunter as 'precise-hunter'} from './modern/classFeatures/ranger/preciseHunter.mjs';
+export {relentlessHunter as 'relentless-hunter'} from './modern/classFeatures/ranger/relentlessHunter.mjs';
 export {tireless} from './modern/classFeatures/ranger/tireless.mjs';
 // Rogue
 export {improvedCunningStrike as 'improved-cunning-strike'} from './modern/classFeatures/rogue/improvedCunningStrike.mjs';
 export {sneakAttack as 'sneak-attack'} from './modern/classFeatures/rogue/sneakAttack.mjs';
 // Sanguine Thief
-export {stolenPower as 'stolen-power'} from './modern/classFeatures/rogue/sanguineThief/stolenPower.mjs';
 export {stealBlood as 'steal-blood'} from './modern/classFeatures/rogue/sanguineThief/stealBlood.mjs';
+export {stolenPower as 'stolen-power'} from './modern/classFeatures/rogue/sanguineThief/stolenPower.mjs';
 // Warlock
 export {stepsOfTheFey as 'steps-of-the-fey'} from './modern/classFeatures/warlock/archfey/stepsOfTheFey.mjs';
 export {felineForm as 'feline-form', felineFormActive as 'feline-form-active', felineFormRevert as 'feline-form-revert'} from './modern/classFeatures/warlock/feline/felineForm.mjs';
-export {createThrall as 'create-thrall', thrallBonus as 'thrall-bonus'} from './modern/classFeatures/warlock/greatOldOne/createThrall.mjs';
 export {awakenedMind as 'awakened-mind'} from './modern/classFeatures/warlock/greatOldOne/awakenedMind.mjs';
+export {createThrall as 'create-thrall', thrallBonus as 'thrall-bonus'} from './modern/classFeatures/warlock/greatOldOne/createThrall.mjs';
 export {psychicSpells as 'psychic-spells'} from './modern/classFeatures/warlock/greatOldOne/psychicSpells.mjs';
 export {thoughtShield as 'thought-shield'} from './modern/classFeatures/warlock/greatOldOne/thoughtShield.mjs';
 export {agonizingBlast as 'agonizing-blast'} from './modern/classFeatures/warlock/invocations/agonizingBlast.mjs';
@@ -63,19 +71,21 @@ export {memorizeSpell as 'memorize-spell'} from './modern/classFeatures/wizard/m
 // ** Spells
 export {chromaticOrb as 'chromatic-orb'} from './modern/spells/chromaticOrb.mjs';
 export {crimsonLash as 'crimson-lash'} from './modern/spells/crimsonLash.mjs';
+export {huntersMark as 'hunters-mark', huntersMarkAttack as 'hunters-mark-attack'} from './modern/spells/huntersMark.mjs';
+export {summonAberration as 'summon-aberration'} from './modern/spells/summonAberration.mjs';
+export {summonFey as 'summon-fey', summonFeyFeyStep as 'summon-fey-fey-step'} from './modern/spells/summonFey.mjs';
 // ** Features
 // Fighting Styles
 export {protectionMoved} from './modern/feats/protection.mjs';
-export {summonAberration as 'summon-aberration'} from './modern/spells/summonAberration.mjs';
 // ** Species Features
 // Grim Hollow Heritage Traits
-export {firstStrike as 'first-strike'} from './modern/speciesFeatures/grimHollowHeritageTraits/firstStrike.mjs';
-export {restorativeRest as 'restorative-rest'} from './modern/speciesFeatures/grimHollowHeritageTraits/restorativeRest.mjs';
-export {powerNap as 'power-nap'} from './modern/speciesFeatures/grimHollowHeritageTraits/powerNap.mjs';
-export {extremeResilience as 'extreme-resilience'} from './modern/speciesFeatures/grimHollowHeritageTraits/extremeResilience.mjs';
 export {etherealFade as 'ethereal-fade', etherealFadeEnd} from './modern/speciesFeatures/grimHollowHeritageTraits/etherealFade.mjs';
+export {extremeResilience as 'extreme-resilience'} from './modern/speciesFeatures/grimHollowHeritageTraits/extremeResilience.mjs';
+export {firstStrike as 'first-strike'} from './modern/speciesFeatures/grimHollowHeritageTraits/firstStrike.mjs';
+export {powerNap as 'power-nap'} from './modern/speciesFeatures/grimHollowHeritageTraits/powerNap.mjs';
+export {restorativeRest as 'restorative-rest'} from './modern/speciesFeatures/grimHollowHeritageTraits/restorativeRest.mjs';
 // ** Dark Bargains
-export {handOfDeath as 'hand-of-death', handOfDeathDeathKnell, handOfDeathClingToLife} from './modern/otherFeatures/darkBargain/handOfDeath.mjs';
+export {handOfDeath as 'hand-of-death', handOfDeathClingToLife, handOfDeathDeathKnell} from './modern/otherFeatures/darkBargain/handOfDeath.mjs';
 // ** Monster Features
 export {shieldGuardianBound as 'bound|shield-guardian', shieldGuardianBoundBind, shieldGuardianBoundDamage} from './modern/monsters/construct/shieldGuardian/bound.mjs';
 // ** Misc